+++ src/zExceptions/__init__.py
@@ -105,12 +105,46 @@ status_reasons = {
 }
 
 
+ERROR_HTML = """\
+<!DOCTYPE html>
+<html>
+<head>
+<title>Site Error</title>
+<meta charset="utf-8" />
+</head>
+<body bgcolor="#FFFFFF">
+<h2>Site Error</h2>
+<p>An error was encountered while publishing this resource.
+</p>
+<p><strong>{title}</strong></p>
+
+{detail}
+<hr noshade="noshade"/>
+
+<p>Troubleshooting Suggestions</p>
+
+<ul>
+<li>The URL may be incorrect.</li>
+<li>The parameters passed to this resource may be incorrect.</li>
+<li>A resource that this resource relies on may be
+  encountering an error.</li>
+</ul>
+
+<p>If the error persists please contact the site maintainer.
+Thank you for your patience.
+</p>
+</body></html>"""
+
+
 @implementer(IHTTPException)
 class HTTPException(Exception):
     body = None
+    body_template = ERROR_HTML
+    detail = None
     empty_body = False
     errmsg = 'Internal Server Error'
     status = 500
+    title = 'Sorry, a site error occurred.'
 
     def setBody(self, body):
         self.body = body
@@ -143,7 +177,11 @@ class HTTPException(Exception):
 
         body = self.body
         if body is None:
-            body = str(self)
+            if self.title and self.detail:
+                body = self.body_template.format(
+                    title=self.title, detail=self.detail)
+            else:
+                body = str(self)
         return [body]